requirements.txt
Django==4.2.7
Pillow==10.1.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
python-decouple==3.8 
cloudinary==1.44.1
django-cloudinary-storage==0.3.0
resend==2.11.0
google-api-python-client==2.141.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
django-ipware==7.0.1
geoip2==5.1.0
